+++ SimulationEngine/src/main/kotlin/wiring/Output.kt
@@ -15,7 +15,7 @@ class Output(id : Int,
              override val componentType: ComponentType = ComponentType.OUTPUT,
              override var highDuration: Int? = null,
              override var lowDuration: Int? = null,
-             override var delayTicks: Int? = null
+             override var delayTicks: Float? = null
     ): BasicComponent(id) {
     override suspend fun evaluate(): MutableList<Pin> {
         return inputs.toMutableList()
