+++ shared/src/commonMain/kotlin/com/example/digisim/LogicGates/Not.kt
@@ -22,7 +22,7 @@ import com.example.digisim.ParsingLogic.ComponentType
      override var height: Float by mutableStateOf(60f)
      override var inputCount: Int by mutableStateOf(initialInputCount)
      override var outputCount: Int by mutableStateOf(initialOutputCount)
-     override var delay: Int? = null
+     override var delay: Float? = null
 
      override val componentType = ComponentType.NOT
